requirements.txt
fastapi
uvicorn
openai
requests
SpeechRecognition
sounddevice
gtts
python-dotenv
pyaudio
numpy
pygame
pycaw